+++ sdk/src/DocuSign.eSign/Model/Group.cs
@@ -41,10 +41,11 @@ namespace DocuSign.eSign.Model
         /// <param name="GroupId">The DocuSign group ID for the group..</param>
         /// <param name="GroupName">The name of the group..</param>
         /// <param name="GroupType">The group type..</param>
+        /// <param name="LastModifiedOn">LastModifiedOn.</param>
         /// <param name="PermissionProfileId">The ID of the permission profile associated with the group..</param>
         /// <param name="Users">Users.</param>
         /// <param name="UsersCount">UsersCount.</param>
-        public Group(string AccessType = default(string), string DsGroupId = default(string), ErrorDetails ErrorDetails = default(ErrorDetails), string GroupId = default(string), string GroupName = default(string), string GroupType = default(string), string PermissionProfileId = default(string), List<UserInfo> Users = default(List<UserInfo>), string UsersCount = default(string))
+        public Group(string AccessType = default(string), string DsGroupId = default(string), ErrorDetails ErrorDetails = default(ErrorDetails), string GroupId = default(string), string GroupName = default(string), string GroupType = default(string), string LastModifiedOn = default(string), string PermissionProfileId = default(string), List<UserInfo> Users = default(List<UserInfo>), string UsersCount = default(string))
         {
             this.AccessType = AccessType;
             this.DsGroupId = DsGroupId;
@@ -52,6 +53,7 @@ namespace DocuSign.eSign.Model
             this.GroupId = GroupId;
             this.GroupName = GroupName;
             this.GroupType = GroupType;
+            this.LastModifiedOn = LastModifiedOn;
             this.PermissionProfileId = PermissionProfileId;
             this.Users = Users;
             this.UsersCount = UsersCount;
@@ -92,6 +94,11 @@ namespace DocuSign.eSign.Model
         [DataMember(Name="groupType", EmitDefaultValue=false)]
         public string GroupType { get; set; }
         /// <summary>
+        /// Gets or Sets LastModifiedOn
+        /// </summary>
+        [DataMember(Name="lastModifiedOn", EmitDefaultValue=false)]
+        public string LastModifiedOn { get; set; }
+        /// <summary>
         /// The ID of the permission profile associated with the group.
         /// </summary>
         /// <value>The ID of the permission profile associated with the group.</value>
@@ -121,6 +128,7 @@ namespace DocuSign.eSign.Model
             sb.Append("  GroupId: ").Append(GroupId).Append("\n");
             sb.Append("  GroupName: ").Append(GroupName).Append("\n");
             sb.Append("  GroupType: ").Append(GroupType).Append("\n");
+            sb.Append("  LastModifiedOn: ").Append(LastModifiedOn).Append("\n");
             sb.Append("  PermissionProfileId: ").Append(PermissionProfileId).Append("\n");
             sb.Append("  Users: ").Append(Users).Append("\n");
             sb.Append("  UsersCount: ").Append(UsersCount).Append("\n");
@@ -191,6 +199,11 @@ namespace DocuSign.eSign.Model
                     this.GroupType.Equals(other.GroupType)
                 ) && 
                 (
+                    this.LastModifiedOn == other.LastModifiedOn ||
+                    this.LastModifiedOn != null &&
+                    this.LastModifiedOn.Equals(other.LastModifiedOn)
+                ) && 
+                (
                     this.PermissionProfileId == other.PermissionProfileId ||
                     this.PermissionProfileId != null &&
                     this.PermissionProfileId.Equals(other.PermissionProfileId)
@@ -230,6 +243,8 @@ namespace DocuSign.eSign.Model
                     hash = hash * 59 + this.GroupName.GetHashCode();
                 if (this.GroupType != null)
                     hash = hash * 59 + this.GroupType.GetHashCode();
+                if (this.LastModifiedOn != null)
+                    hash = hash * 59 + this.LastModifiedOn.GetHashCode();
                 if (this.PermissionProfileId != null)
                     hash = hash * 59 + this.PermissionProfileId.GetHashCode();
                 if (this.Users != null)