+++ test/SdkNetCoreTests/JwtAuthNetCoreUnitTests.cs
@@ -320,7 +320,7 @@ namespace SdkNetCoreTests
         {
             JwtRequestSignatureOnDocumentTest("created");
 
-            ReturnUrlRequest options = new ReturnUrlRequest();
+            EnvelopeViewRequest options = new EnvelopeViewRequest();
             options.ReturnUrl = _testConfig.ReturnUrl;
 
             // |EnvelopesApi| contains methods related to envelopes and envelope recipients